move appActivate
